+++ ChangeLog.md
@@ -3,6 +3,7 @@
   * Enable update and outdate messages
   * Ensure that VoiceAttack decimal values are not written as integers
   * Add 'SetState' Cottle function.  Details on how to use this are in the SpeechResponder documentation
+  * Allow speech responder to not speak when subtitles are written
 
 #2.1.0-b1
   * Fix material name for Cadmium